+++ tests/Adapter/LoggerTest.php
@@ -5,8 +5,6 @@ namespace Tests\Innmind\Filesystem\Adapter;
 
 use Innmind\Filesystem\{
     Adapter,
-    Adapter\Logger,
-    Adapter\InMemory,
     File,
     File\Content,
     Name,
@@ -21,8 +19,8 @@ class LoggerTest extends TestCase
     {
         $this->assertInstanceOf(
             Adapter::class,
-            Logger::psr(
-                InMemory::emulateFilesystem(),
+            Adapter::logger(
+                Adapter::inMemory(),
                 new NullLogger,
             ),
         );
@@ -30,8 +28,8 @@ class LoggerTest extends TestCase
 
     public function testAdd()
     {
-        $adapter = Logger::psr(
-            $inner = InMemory::emulateFilesystem(),
+        $adapter = Adapter::logger(
+            $inner = Adapter::inMemory(),
             new NullLogger,
         );
         $file = File::of(Name::of('foo'), Content::none());
@@ -47,8 +45,8 @@ class LoggerTest extends TestCase
 
     public function testGet()
     {
-        $adapter = Logger::psr(
-            $inner = InMemory::emulateFilesystem(),
+        $adapter = Adapter::logger(
+            $inner = Adapter::inMemory(),
             new NullLogger,
         );
         $name = Name::of('foo');
@@ -68,8 +66,8 @@ class LoggerTest extends TestCase
 
     public function testContains()
     {
-        $adapter = Logger::psr(
-            $inner = InMemory::emulateFilesystem(),
+        $adapter = Adapter::logger(
+            $inner = Adapter::inMemory(),
             new NullLogger,
         );
         $name = Name::of('foo');
@@ -82,8 +80,8 @@ class LoggerTest extends TestCase
 
     public function testRemove()
     {
-        $adapter = Logger::psr(
-            $inner = InMemory::emulateFilesystem(),
+        $adapter = Adapter::logger(
+            $inner = Adapter::inMemory(),
             new NullLogger,
         );
         $name = Name::of('foo');
@@ -102,8 +100,8 @@ class LoggerTest extends TestCase
 
     public function testRoot()
     {
-        $adapter = Logger::psr(
-            $inner = InMemory::emulateFilesystem(),
+        $adapter = Adapter::logger(
+            $inner = Adapter::inMemory(),
             new NullLogger,
         );
         $file = File::named(