+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Links are not silently ignored when listing files/directories. An error is still returned when trying to remove a link.
 - `Innmind\Filesystem\Adapter\Filesystem::mount()` now returns an `Innmind\Immutable\Attempt`
 - `Innmind\Filesystem\Directory::removed()` is now flagged as internal
+- `Innmind\Filesystem\Adapter` is now a final class
+- `Innmind\Filesystem\Adapter\Filesystem` is now flagged as internal
+- `Innmind\Filesystem\Adapter\InMemory` is now flagged as internal
+- `Innmind\Filesystem\Adapter\Logger` is now flagged as internal
 
 ### Removed